# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Supabase Auth Express API",
    "version": "1.0.0",
    "description": "Express.js API implementing Supabase Auth"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "User Signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns access_token and refresh_token"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid login credentials"
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Get Public Info",
        "responses": {
          "200": {
            "description": "Public info message"
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get Protected User Profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile with id, email, and created_at"
          },
          "401": {
            "description": "Access token required or Invalid/expired token"
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get Protected Dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard data with welcome message and user info"
          },
          "401": {
            "description": "Access token required or Invalid/expired token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout User",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logout successful, no body"
          },
          "401": {
            "description": "Access token required or Invalid/expired token"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}